update colormaps to changes in viscm color space
viscm had a bug in its color space model. This updates the color maps to
the new, corrected model. However, the model change alone would have
increased the saturation. To compensate for this, the colormap
parameters were altered slightly to give the same appearance in the new
color space model.